modification login (non finalisé)

+++ testClimatCookie.php
@@ -1,18 +1,32 @@
 <?php
 session_start();
-include('navigation/bdconnexion.php');
-if ($_POST['suppDef'] == 'suppDef') {
-    $user = strip_tags($_COOKIE['user']);
-    $sqlQuery = "SHOW TABLES FROM testclimat LIKE '" . $user . "'";
-    $verifTable = $db->prepare($sqlQuery);
-    $verifTable->execute();
-    $table = $verifTable->fetchAll();
-    if ($table != null) {
-        $deleteUserDataBase = $db->prepare('DROP TABLE ' . $user);
-        $deleteUserDataBase->execute([]) or die(print_r($db->errorInfo()));
-    }
-    unset($_COOKIE['user']);
+include('connexion/bdconnexion.php');
+$USERclef = strip_tags($_COOKIE['logged']) ?? null;
+if ($USERclef == null) {
+    header('Location: userThingsLogin.php');
+    exit();
 }
+$sqlQuery = 'SELECT * FROM user WHERE clef = :clef';
+$verifUser = $db->prepare($sqlQuery);
+$verifUser->execute(['clef' => $USERclef]);
+$user = $verifUser->fetch();
+
+if (isset($_POST['suppDef'])) {
+    if ($_POST['suppDef'] == 'suppDef') {
+        $user = strip_tags($_COOKIE['user']);
+        $sqlQuery = "SHOW TABLES FROM testclimat LIKE '" . $user . "'";
+        $verifTable = $db->prepare($sqlQuery);
+        $verifTable->execute();
+        $table = $verifTable->fetchAll();
+        if ($table != null) {
+            $deleteUserDataBase = $db->prepare('DROP TABLE ' . $user);
+            $deleteUserDataBase->execute([]) or die(print_r($db->errorInfo()));
+        }
+        unset($_COOKIE['user']);
+        unset($_COOKIE['logged']);
+    }
+}
+
 ?>
 <!DOCTYPE html>
 <html>
@@ -27,34 +41,29 @@ if ($_POST['suppDef'] == 'suppDef') {
     <?php include('navigation/nav.php'); ?>
     <main>
         <section class="section_intro">
-            <h1>Gestion du cookie</h1>
-            <h2>Présentation</h2>
-            <p>Le cookie est un petit fichier texte stocké dans votre navigateur internet. <br />Dans le cas de testClimat, le cookie est utilisé pour stocker un identifiant unique qui permet de vous associer à une base de données. Cette base de données est utilisée pour stocker les climats que vous avez sauvegardé. Par conséquent, si vous avez désactivé les cookies, vous ne pourrez pas sauvegarder vos climats.</p>
-            <p>
-                Votre cookie : <?php if (isset($_COOKIE['user'])) {
-                                    echo '<div class="flou"> ' . $_COOKIE["user"] . ' </div> (passez votre souris au dessus pour le voir).';
-                                }
-                                if (isset($_COOKIE['logged'])) {
-                                    echo '<div class="flou"> ' . $_COOKIE["logged"] . ' </div> (passez votre souris au dessus pour le voir).';
-                                } else {
-                                    echo 'Aucun cookie ne vous est assigné.';
-                                    echo '<br />Aucune base de données ne vous est assigné.';
-                                }
-                                ?>
+            <h1>Gestion du compte</h1>
+            <h2>Information sur votre compte</h2>
+            <p>Vous êtes connecté en tant que :
+                <?php
+                echo $user['mail'];
+                echo '<br />Votre date d\'inscription est : ' . $user['dateInscription'];
+                echo '<br />Votre mail est vérifié : ';
+                $mailVERIF = $user['mailVerifie'] ? '1' : 'non';
+                echo $mailVERIF;
+                ?>
             </p>
         </section>
         <section class="section_milieu">
-            <h2>Suppression du cookie</h2>
-            <p>Pensez à refuser l'accès au cookie à testClimat sans quoi celui en recréra un à votre prochaine connexion. <br /><b> Le bouton ci-dessous effacera votre cookie et la base de données associée. </b><br />En effet, une base de données au nom de votre cookie se créer automatiquement à la première utilisation de testClimat (en mode avec cookie).<br /><br />
-            <div class="A_noter">
-                <div class="A_noter_titre">A noter</div>
-                Si vous réutilisez testClimat par la suite sans avoir désactivé les cookies pour le site, une nouvelle table dans la base de données vous sera associée.
-            </div>
+            <h2>Suppression du compte</h2>
+            <h3>Présentation des cookies</h3>
+            <p>Le cookie est un petit fichier texte stocké dans votre navigateur internet. <br />Dans le cas de testClimat, le cookie est utilisé pour stocker un identifiant unique qui permet de vous associer à vos normales climatiques dans un base de données. Cette base de données est utilisée pour stocker les climats que vous avez sauvegardé. Par conséquent, si vous avez désactivé les cookies, vous ne pourrez pas sauvegarder vos climats.</p>
+            <p><b> Le bouton ci-dessous effacera votre compte et vos données climatiques enregistré </b></p>
         </section>
         <section class="section_fin">
             <div class="attention">
                 <div class="attention_titre">Attention</div>
                 <p><b>Toutes les données effacées seront irrémédiablement effacées</b><br />Nous vous conseillons de telécharger vos climats enregistrés.</p>
+
                 <form method='post'>
                     <button method='post' type="submit" name="supp" value="supp">Supprimer mon cookie et la base de données associée</button>
                 </form>
@@ -62,15 +71,14 @@ if ($_POST['suppDef'] == 'suppDef') {
                     <?php
                     if (isset($_POST['supp'])) {
                         if ($_POST['supp'] == 'supp') {
-                            echo '<br /><button method="post" type="submit" name="suppDef" value="suppDef">Confirmer la suppréssion de mon cookie et de la base de données associée</button>';
+                            echo '<br /><button method="post" type="submit" name="suppDef" value="suppDef">Confirmer la suppréssion de mon compte</button>';
                         }
                         if (isset($_POST['suppDef'])) {
                             if ($_POST['suppDef'] == 'suppDef') {
-                                echo 'Votre cookie ainsi que votre base de données ont été supprimé <br /><br />Nous vous invitons à fermer cette fenêtre.<br /><a href="index.php">Pour revenir à l\'accueil  de testClimat cliquez ici</a><br /><b>Attention, si vous avez pas refusé l\'accès au cookie à testClimat depuis votre navigateur, un nouveau cookie se régénéra si vous cliquer sur le lien</b>';
+                                echo 'Votre compte ainsi que toutes vos données ont été supprimé <br /><br />Nous vous invitons à fermer cette page.<br /><a href="index.php">Pour revenir à l\'accueil  de testClimat cliquez ici</a>';
                             }
                         }
                     }
-
                     ?>
                 </form>
             </div>
